@@ -4,12 +4,12 @@
   <parent>
     <groupId>com.amazon.alexa</groupId>
     <artifactId>ask-sdk-pom</artifactId>
-    <version>2.86.0</version>
+    <version>2.86.1-jdk17</version>
   </parent>
   <groupId>com.amazon.alexa</groupId>
   <artifactId>ask-sdk-apache-client</artifactId>
   <packaging>jar</packaging>
-  <version>2.86.0</version>
+  <version>2.86.1-jdk17</version>
   <name>ASK SDK for Java Apache API Client</name>
   <description>Apache HTTP Client API client plugin for the ASK SDK for Java.</description>
   <url>http://developer.amazon.com/ask</url>
@@ -68,8 +68,8 @@
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.7.0</version>
           <configuration>
-            <source>1.8</source>
-            <target>1.8</target>
+            <source>17</source>
+            <target>17</target>
             <forceJavacCompilerUse>true</forceJavacCompilerUse>
           </configuration>
         </plugin>
