@@ -4,12 +4,12 @@
     <parent>
         <groupId>com.amazon.alexa</groupId>
         <artifactId>ask-sdk-pom</artifactId>
-        <version>2.86.0</version>
+        <version>2.86.1-jdk17</version>
     </parent>
     <groupId>com.amazon.alexa</groupId>
     <artifactId>ask-sdk-freemarker</artifactId>
     <packaging>jar</packaging>
-    <version>2.86.0</version>
+    <version>2.86.1-jdk17</version>
     <name>ASK SDK for Java FreeMarker Template Renderer</name>
     <description>Template Renderer for ASK SDK Java using the Apache Freemarker template engine.</description>
     <url>http://developer.amazon.com/ask</url>
@@ -37,7 +37,7 @@
         <dependency>
             <groupId>com.amazon.alexa</groupId>
             <artifactId>ask-sdk-runtime</artifactId>
-            <version>2.86.0</version>
+            <version>2.86.1-jdk17</version>
         </dependency>
         <dependency>
             <groupId>org.freemarker</groupId>
@@ -73,8 +73,8 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.7.0</version>
                     <configuration>
-                        <source>1.8</source>
-                        <target>1.8</target>
+                        <source>17</source>
+                        <target>17</target>
                         <forceJavacCompilerUse>true</forceJavacCompilerUse>
                     </configuration>
                 </plugin>
