@@ -4,12 +4,12 @@
     <parent>
         <groupId>com.amazon.alexa</groupId>
         <artifactId>ask-sdk-pom</artifactId>
-        <version>2.86.0</version>
+        <version>2.86.1-jdk17</version>
     </parent>
     <groupId>com.amazon.alexa</groupId>
     <artifactId>ask-sdk-local-debug</artifactId>
     <packaging>jar</packaging>
-    <version>2.86.0</version>
+    <version>2.86.1-jdk17</version>
     <name>ASK SDK for Java Local Debugging Adapter</name>
     <description>Enables local debugging support for skills.</description>
     <url>http://developer.amazon.com/ask</url>
@@ -37,12 +37,12 @@
         <dependency>
             <groupId>com.amazon.alexa</groupId>
             <artifactId>ask-sdk-lambda-support</artifactId>
-            <version>2.86.0</version>
+            <version>2.86.1-jdk17</version>
         </dependency>
         <dependency>
             <groupId>com.amazon.alexa</groupId>
             <artifactId>ask-sdk-apache-client</artifactId>
-            <version>2.86.0</version>
+            <version>2.86.1-jdk17</version>
         </dependency>
         <dependency>
             <groupId>com.amazon.alexa</groupId>
@@ -99,8 +99,8 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.7.0</version>
                     <configuration>
-                        <source>1.8</source>
-                        <target>1.8</target>
+                        <source>17</source>
+                        <target>17</target>
                         <forceJavacCompilerUse>true</forceJavacCompilerUse>
                     </configuration>
                 </plugin>
