@@ -4,12 +4,12 @@
   <parent>
     <groupId>com.amazon.alexa</groupId>
     <artifactId>ask-sdk-pom</artifactId>
-    <version>2.86.0</version>
+    <version>2.86.1-jdk17</version>
   </parent>
   <groupId>com.amazon.alexa</groupId>
   <artifactId>ask-sdk-servlet-support</artifactId>
   <packaging>jar</packaging>
-  <version>2.86.0</version>
+  <version>2.86.1-jdk17</version>
   <name>ASK SDK for Java Servlet Support Plugin</name>
   <description>Contains supporting components for hosting a skill on a Servlet with the ASK SDK for Java.</description>
   <url>http://developer.amazon.com/ask</url>
@@ -37,14 +37,16 @@
     <dependency>
       <groupId>com.amazon.alexa</groupId>
       <artifactId>ask-sdk-core</artifactId>
-      <version>2.86.0</version>
+      <version>2.86.1-jdk17</version>
     </dependency>
+
     <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>3.0.1</version>
-      <scope>provided</scope>
+        <groupId>jakarta.servlet</groupId>
+        <artifactId>jakarta.servlet-api</artifactId>
+        <version>5.0.0</version>
+        <scope>provided</scope>
     </dependency>
+
     <dependency>
       <groupId>commons-codec</groupId>
       <artifactId>commons-codec</artifactId>
@@ -102,8 +104,8 @@
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.7.0</version>
           <configuration>
-            <source>1.8</source>
-            <target>1.8</target>
+            <source>17</source>
+            <target>17</target>
             <forceJavacCompilerUse>true</forceJavacCompilerUse>
           </configuration>
         </plugin>
